Moved wayfinder library to dvl_wayfinder_lib
